Update footer to shared src with primaryColor set to wrapping td
Remove footerLogoSrc with this change to how it renders  the footer logo.  It no longer uses the color specific ones.  Now it will use a shared logo with a white upper right corner mask and have the bgcolor set to the primaryColor.  This achieves the same this while matching footer colors in dark mode.  The only remaining issue is that the upper right corner reamins white instead of fliping to black or what ever color dark mode should be.